fix: 修复 int 转 Long 类型不匹配问题

+++ backend/src/main/java/com/imagemanager/service/impl/DashboardServiceImpl.java
@@ -115,7 +115,7 @@ public class DashboardServiceImpl implements DashboardService {
             totalAlbums,
             totalTags,
             favoritesCount,
-            trashImages.size(),
+            (long) trashImages.size(),
             recentUploads7d,
             recentUploads30d
         );
